// apps/portal/app/interfaces/tools_interfaces/IConversationOptions.ts
export interface IConversationOptions {
  filename: string;
  category: string;
  data: string;
  fillColor: string;
  strokeColor: string;
  strokeWidth: number;
  textColor: string;
  textStrokeColor: string;
  textStrokeWidth: number;
  textSize: number;
  fontFamily: string;
  dash: number[];
}